api/requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
flask-cors==3.0.10
requests==2.26.0
bs4==0.0.1
beautifulsoup4==4.10.0
python-dotenv==0.19.2
pykakasi==2.2.1
flask-limiter==3.3.1
limits==3.5.0
